refactor(diagnostics): keep issue-draft composition out of the background
The background bundle imported `getSafeClientEnvironment` from `error-report`,
which pulled the whole module — issue templates, URL length budgeting, the
privacy tail, every draft builder — into a chunk that has a size budget and no
use for any of it. Adding the migration summary to the draft tipped it 9 bytes
past the 193000-byte ceiling, which is how the coupling surfaced.

The environment helper moves to its own module. `error-report` re-exports it, so
existing importers are unchanged.

Background gzip: 193009 -> 192540, now under budget with the new diagnostics
code rather than by raising the ceiling for it.
